Time: 1 ms (100.00%) | Memory: 42.4 MB (90.60%) - LeetSync

+++ 4168-mirror-distance-of-an-integer/mirror-distance-of-an-integer.java
@@ -1,14 +1,14 @@
 class Solution {
     public int mirrorDistance(int n) {
-        return (int)Math.abs(n - rev(n));
+        return (int) Math.abs(n - rev(n));
     }
 
-    private long rev(int n) {
+    private int rev(int n) {
         long sum = 0;
         while (n > 0) {
             sum = sum * 10 + n % 10;
             n /= 10;
         }
-        return sum;
+        return (int) sum;
     }
 }
